Dump via api on windows
