+++ src/Microsoft.TestPlatform.Extensions.BlameDataCollector/DotnetProcessDumpUtility.cs
@@ -1,25 +1,25 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
-# if NETSTANDARD
+//# if NETSTANDARD
 
 namespace Microsoft.TestPlatform.Extensions.BlameDataCollector
 {
     using System;
     using System.Diagnostics;
     using System.IO;
-    using System.Runtime.InteropServices;
-    using Microsoft.Diagnostics.NETCore.Client;
     using Microsoft.VisualStudio.TestPlatform.ObjectModel;
     using Microsoft.VisualStudio.TestPlatform.PlatformAbstractions;
     using Microsoft.VisualStudio.TestPlatform.PlatformAbstractions.Interfaces;
     using Microsoft.VisualStudio.TestPlatform.Utilities.Helpers;
     using Microsoft.VisualStudio.TestPlatform.Utilities.Helpers.Interfaces;
-    using Microsoft.Win32.SafeHandles;
-
+#if NETSTANDARD
+    using System.Runtime.InteropServices;
+    using Microsoft.Diagnostics.NETCore.Client;
+#endif
     internal class DotnetProcessDumpUtility : IProcessDumpUtility
     {
-        private IProcessHelper processHelper;
-        private IFileHelper fileHelper;
+        private readonly IProcessHelper processHelper;
+        private readonly IFileHelper fileHelper;
         private string dumpPath;
 
         public DotnetProcessDumpUtility()
@@ -100,7 +100,7 @@ namespace Microsoft.TestPlatform.Extensions.BlameDataCollector
             // this is just nice to have info, should we continue if this fails? --jajares
             var processName = this.processHelper.GetProcessName(processId);
 
-            var dumpType = isFullDump ? DumpType.Full : DumpType.Normal;
+            var dumpType = isFullDump ? DumpTypeOption.Full : DumpTypeOption.Mini;
             // the below format is extremely ugly maybe we can use: 
             // https://github.com/microsoft/testfx/issues/678
             // which will order the files correctly gives more info when transported out of 
@@ -124,27 +124,31 @@ namespace Microsoft.TestPlatform.Extensions.BlameDataCollector
             }
 
             var dumpPath = Path.Combine(path, dumpFileName);
-            var client = new DiagnosticsClient(processId);
+           
             EqtTrace.Info($"DotnetProcessDumpUtility.StartHangBasedProcessDump: Creating {dumpType.ToString().ToLowerInvariant()} dump of process {processName} ({processId}) into '{dumpPath}'.");
             this.dumpPath = dumpPath;
 
             try
             {
+#if NETSTANDARD
                 if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
                 {
+#endif
                     // Get the process
                     Process process = Process.GetProcessById(processId);
                     DumperWindows.CollectDumpAsync(process, dumpPath, dumpType);
+#if NETSTANDARD
                 }
                 else if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
                 {
-
-                    client.WriteDump(DumpType.Full, dumpPath);
+                    var client = new DiagnosticsClient(processId);
+                    client.WriteDump(dumpType == DumpTypeOption.Full ? DumpType.Full: DumpType.Normal, dumpPath);
                 }
                 else
                 {
                     throw new PlatformNotSupportedException($"Unsupported operating system: {RuntimeInformation.OSDescription}");
                 }
+#endif
             }
             catch (Exception ex)
             {
@@ -166,111 +170,7 @@ namespace Microsoft.TestPlatform.Extensions.BlameDataCollector
         {
             // noop
         }
-
-        private static class DumperWindows
-        {
-            internal static void CollectDumpAsync(Process process, string outputFile, DumpType type)
-            {
-                // We can't do this "asynchronously" so just Task.Run it. It shouldn't be "long-running" so this is fairly safe.
-
-                // Open the file for writing
-                using (var stream = new FileStream(outputFile, FileMode.Create, FileAccess.ReadWrite, FileShare.None))
-                {
-                    var exceptionInfo = new NativeMethods.MINIDUMP_EXCEPTION_INFORMATION();
-
-                    NativeMethods.MINIDUMP_TYPE dumpType = NativeMethods.MINIDUMP_TYPE.MiniDumpNormal;
-                    switch (type)
-                    {
-                        case DumpType.Full:
-                            dumpType = NativeMethods.MINIDUMP_TYPE.MiniDumpWithFullMemory |
-                                       NativeMethods.MINIDUMP_TYPE.MiniDumpWithDataSegs |
-                                       NativeMethods.MINIDUMP_TYPE.MiniDumpWithHandleData |
-                                       NativeMethods.MINIDUMP_TYPE.MiniDumpWithUnloadedModules |
-                                       NativeMethods.MINIDUMP_TYPE.MiniDumpWithFullMemoryInfo |
-                                       NativeMethods.MINIDUMP_TYPE.MiniDumpWithThreadInfo |
-                                       NativeMethods.MINIDUMP_TYPE.MiniDumpWithTokenInformation;
-                            break;
-                        case DumpType.WithHeap:
-                            dumpType = NativeMethods.MINIDUMP_TYPE.MiniDumpWithPrivateReadWriteMemory |
-                                       NativeMethods.MINIDUMP_TYPE.MiniDumpWithDataSegs |
-                                       NativeMethods.MINIDUMP_TYPE.MiniDumpWithHandleData |
-                                       NativeMethods.MINIDUMP_TYPE.MiniDumpWithUnloadedModules |
-                                       NativeMethods.MINIDUMP_TYPE.MiniDumpWithFullMemoryInfo |
-                                       NativeMethods.MINIDUMP_TYPE.MiniDumpWithThreadInfo |
-                                       NativeMethods.MINIDUMP_TYPE.MiniDumpWithTokenInformation;
-                            break;
-                        case DumpType.Normal:
-                            dumpType = NativeMethods.MINIDUMP_TYPE.MiniDumpWithThreadInfo;
-                            break;
-                    }
-
-                    // Retry the write dump on ERROR_PARTIAL_COPY
-                    for (int i = 0; i < 5; i++)
-                    {
-                        // Dump the process!
-                        if (NativeMethods.MiniDumpWriteDump(process.Handle, (uint)process.Id, stream.SafeFileHandle, dumpType, ref exceptionInfo, IntPtr.Zero, IntPtr.Zero))
-                        {
-                            break;
-                        }
-                        else
-                        {
-                            int err = Marshal.GetHRForLastWin32Error();
-                            if (err != NativeMethods.ERROR_PARTIAL_COPY)
-                            {
-                                Marshal.ThrowExceptionForHR(err);
-                            }
-                        }
-                    }
-                }
-            }
-
-            private static class NativeMethods
-            {
-                public const int ERROR_PARTIAL_COPY = unchecked((int)0x8007012b);
-
-                [DllImport("Dbghelp.dll", SetLastError = true)]
-                public static extern bool MiniDumpWriteDump(IntPtr hProcess, uint ProcessId, SafeFileHandle hFile, MINIDUMP_TYPE DumpType, ref MINIDUMP_EXCEPTION_INFORMATION ExceptionParam, IntPtr UserStreamParam, IntPtr CallbackParam);
-
-                [StructLayout(LayoutKind.Sequential, Pack = 4)]
-                public struct MINIDUMP_EXCEPTION_INFORMATION
-                {
-                    public uint ThreadId;
-                    public IntPtr ExceptionPointers;
-                    public int ClientPointers;
-                }
-
-                [Flags]
-                public enum MINIDUMP_TYPE : uint
-                {
-                    MiniDumpNormal = 0,
-                    MiniDumpWithDataSegs = 1 << 0,
-                    MiniDumpWithFullMemory = 1 << 1,
-                    MiniDumpWithHandleData = 1 << 2,
-                    MiniDumpFilterMemory = 1 << 3,
-                    MiniDumpScanMemory = 1 << 4,
-                    MiniDumpWithUnloadedModules = 1 << 5,
-                    MiniDumpWithIndirectlyReferencedMemory = 1 << 6,
-                    MiniDumpFilterModulePaths = 1 << 7,
-                    MiniDumpWithProcessThreadData = 1 << 8,
-                    MiniDumpWithPrivateReadWriteMemory = 1 << 9,
-                    MiniDumpWithoutOptionalData = 1 << 10,
-                    MiniDumpWithFullMemoryInfo = 1 << 11,
-                    MiniDumpWithThreadInfo = 1 << 12,
-                    MiniDumpWithCodeSegs = 1 << 13,
-                    MiniDumpWithoutAuxiliaryState = 1 << 14,
-                    MiniDumpWithFullAuxiliaryState = 1 << 15,
-                    MiniDumpWithPrivateWriteCopyMemory = 1 << 16,
-                    MiniDumpIgnoreInaccessibleMemory = 1 << 17,
-                    MiniDumpWithTokenInformation = 1 << 18,
-                    MiniDumpWithModuleHeaders = 1 << 19,
-                    MiniDumpFilterTriage = 1 << 20,
-                    MiniDumpWithAvxXStateContext = 1 << 21,
-                    MiniDumpWithIptTrace = 1 << 22,
-                    MiniDumpValidTypeFlags = (-1) ^ ((~1) << 22)
-                }
-            }
-        }
     }
 }
 
-#endif
+//#endif
